Header component
Created a header component and added it into the App component.  Header
includes a logo and links to shop and contact page.  Created styling for
the header and also styled the app for 'a' tags so they won't have
text decorations and will always have the text color of black.